Patch 0.1.3: Refactored profession page -> grid -> component
