+++ src/components/subcomponents/PdfViewer.js
@@ -6,10 +6,7 @@ import Grid from "@material-ui/core/Grid";
 import Card from "@material-ui/core/Card";
 
 const useStyles = makeStyles((theme) => ({
-    pdfviewer: {
-        padding: theme.spacing(2),
-        marginTop: theme.spacing(2),
-    },
+
 }));
 
 function PdfViewer(props) {
@@ -23,15 +20,13 @@ function PdfViewer(props) {
     }
 
     return (
-        <div className={classes.pdfviewer}>
-            <Grid container spacing={2} justify='center' alignItems='center'>
-                <Card>
-                    <Document file={src} onLoadSuccess={onDocumentLoadSuccess}>
-                        <Page pageNumber={pageNumber}/>
-                    </Document>
-                </Card>
-            </Grid>
-        </div>
+        <Grid container spacing={2} justify='center' alignItems='center'>
+            <Card>
+                <Document file={src} onLoadSuccess={onDocumentLoadSuccess}>
+                    <Page pageNumber={pageNumber}/>
+                </Document>
+            </Card>
+        </Grid>
     );
 
 }
